+++ tests/service/mod.rs
@@ -24,7 +24,12 @@ pub fn test_in_tempdir<E: Into<failure::Error>>(
     let result = panic::catch_unwind(move || -> Fallible<()> {
         std::fs::write(
             tempdir_path.join("snowchains.toml"),
-            include_bytes!("../snowchains.toml").as_ref(),
+            &include_bytes!("../snowchains.toml")[..],
+        )?;
+        std::fs::create_dir(tempdir_path.join(".snowchains"))?;
+        std::fs::write(
+            tempdir_path.join(".snowchains").join("target.json"),
+            &include_bytes!("../target.json")[..],
         )?;
         std::fs::create_dir(tempdir_path.join("local"))?;
         serde_json::to_writer(
@@ -87,7 +92,13 @@ pub fn confirm_num_cases(
     }
 
     for &(problem, expected_num_cases) in pairs {
-        let path = wd.join(format!("{}/{}/tests/{}.yml", service, contest, problem));
+        let path = wd
+            .join(".snowchains")
+            .join("tests")
+            .join(Into::<&str>::into(service))
+            .join(contest)
+            .join(problem)
+            .with_extension("yml");
         let file = File::open(&path)?;
         let suite = serde_yaml::from_reader::<_, BatchSuite>(file)?;
         assert_eq!(expected_num_cases, suite.cases.len());